+++ s2_fish/macro_bunch_store.h
@@ -31,7 +31,7 @@ struct Macro_bunch_store
 
   Macro_bunch_store(numeric::array& numeric_local_particles,
 	      int local_num, int total_num, 
-          double mass, double charge, double total_current,
+          double mass, int charge, double total_current,
 	      numeric::array& numeric_units,
 	      numeric::array& numeric_ref_particle,
 	      bool is_fixedz);
